refactor: create MapWithLocationsAndPath component (#304)
* refactor: create MapWithLocationsAndPath component

* refactor: create map-related components folder

* refactor: move useAgencyList

* refactor: move map component

* fix: export Path

* fix: bad merge commit

* fix: map expand button

* chore: fix lint
