+++ src/pages/Map.scss
@@ -56,4 +56,6 @@ pre {
   position: absolute;
   top: 0;
   left: 0;
+  bottom: 0;
+  right: 0;
 }
